@@ -1,4 +1,5 @@
 {
+  "type": "module",
   "author": {
     "name": "ShiwaForce",
     "email": "hello@shiwaforce.com"
@@ -6,13 +7,7 @@
   "bugs": {
     "url": "https://github.com/shiwaforce/eslint-config-server-shiwaforce/issues"
   },
-  "dependencies": {
-    "extend": "^3.0.0"
-  },
   "description": "Shareable server side config for eslint by ShiwaForce",
-  "devDependencies": {
-    "eslint": "^4.0.0"
-  },
   "homepage": "https://github.com/shiwaforce/eslint-config-server-shiwaforce#readme",
   "keywords": [
     "eslint",
@@ -40,5 +35,11 @@
   "scripts": {
     "lint": "./node_modules/.bin/eslint index.js rules/*.js"
   },
-  "version": "1.0.6"
+  "version": "1.0.6",
+  "peerDependencies": {
+    "eslint": "^9.1.1"
+  },
+  "dependencies": {
+    "globals": "^15.1.0"
+  }
 }
