@@ -2,14 +2,14 @@
  * beauty of code
  * @author sarkiroka on 2017.02.17.
  */
-module.exports = {
-	'brace-style':                 [2],
-	'camelcase':                   [2],
-	'dot-location':                [2, 'property'],
-	'dot-notation':                [2],
-	'eqeqeq':                      [0, 'smart'],
-	'new-cap':                     [2],
-	'new-parens':                  [2],
-	'quotes':                      [1, 'single'],
-	'yoda':                        [2, 'never']
+export default {
+	'brace-style': ['error'],
+	'camelcase': ['error'],
+	'dot-location': ['error', 'property'],
+	'dot-notation': ['error'],
+	'eqeqeq': ['off', 'smart'],
+	'new-cap': ['error'],
+	'new-parens': ['error'],
+	'quotes': ['error', 'single'],
+	'yoda': ['error', 'never']
 };
