@@ -2,7 +2,7 @@
  * debugging rules
  * @author sarkiroka on 2017.02.17.
  */
-module.exports = {
-	'no-debugger':                 [2],
-	'no-alert':                    [2]
+export default {
+	'no-debugger': ['error'],
+	'no-alert': ['error']
 };
