@@ -2,6 +2,6 @@
  * rule indicated by using git
  * @author sarkiroka on 2017.02.17.
  */
-module.exports = {
-	'eol-last':                    [2]
+export default {
+	'eol-last': ['error']
 };
