@@ -2,11 +2,11 @@
  * separator rules
  * @author sarkiroka on 2017.02.17.
  */
-module.exports = {
-	'comma-dangle':                [2, 'never'],
-	'comma-spacing':               [1],
-	'comma-style':                 [2],
-	'no-extra-semi':               [2],
-	'semi':                        [2, 'always'],
-	'semi-spacing':                [2]
+export default {
+	'comma-dangle': ['error', 'never'],
+	'comma-spacing': ['warn'],
+	'comma-style': ['error'],
+	'no-extra-semi': ['error'],
+	'semi': ['error', 'always'],
+	'semi-spacing': ['error']
 };
