@@ -2,19 +2,19 @@
  * possible errors
  * @author sarkiroka on 2017.02.17.
  */
-module.exports = {
-	'accessor-pairs':              [2],
-	'block-scoped-var':            [2],
-	'consistent-return':           [2],
-	'default-case':                [2],
-	'handle-callback-err':         [2],
-	'no-array-constructor':        [2],
-	'no-bitwise':                  [2],
-	'no-caller':                   [2],
-	'no-cond-assign':              [2],
-	'no-dupe-args':                [2],
-	'no-dupe-keys':                [2],
-	'no-process-exit':             [2],
-	'no-unreachable':              [2],
-	'no-warning-comments':         [1]
+export default {
+	'accessor-pairs': ['error'],
+	'block-scoped-var': ['error'],
+	'consistent-return': ['error'],
+	'default-case': ['error'],
+	'handle-callback-err': ['error'],
+	'no-array-constructor': ['error'],
+	'no-bitwise': ['error'],
+	'no-caller': ['error'],
+	'no-cond-assign': ['error'],
+	'no-dupe-args': ['error'],
+	'no-dupe-keys': ['error'],
+	'no-process-exit': ['error'],
+	'no-unreachable': ['error'],
+	'no-warning-comments': ['warn']
 };
