@@ -2,62 +2,63 @@
  * others
  * @author sarkiroka on 2017.02.17.
  */
-module.exports = {
-	'no-constant-condition':       [2],
-	'no-control-regex':            [2],
-	'no-delete-var':               [2],
-	'no-div-regex':                [2],
-	'no-duplicate-case':           [2],
-	'no-empty-character-class':    [2],
-	'no-ex-assign':                [2],
-	'no-extend-native':            [2],
-	'no-extra-bind':               [2],
-	'no-extra-boolean-cast':       [2],
-	'no-extra-parens':             [2],
-	'no-fallthrough':              [2],
-	'no-floating-decimal':         [2],
-	'no-func-assign':              [2],
-	'no-implied-eval':             [2],
-	'no-inner-declarations':       [2],
-	'no-invalid-regexp':           [2],
-	'no-iterator':                 [2],
-	'no-label-var':                [2],
-	'no-labels':                   [2],
-	'no-multi-str':                [2],
-	'no-multiple-empty-lines':     [2],
-	'no-native-reassign':          [2],
-	'no-negated-in-lhs':           [2],
-	'no-nested-ternary':           [2],
-	'no-new':                      [2],
-	'no-new-func':                 [2],
-	'no-new-object':               [2],
-	'no-new-require':              [2],
-	'no-new-wrappers':             [2],
-	'no-obj-calls':                [2],
-	'no-octal':                    [2],
-	'no-octal-escape':             [2],
-	'no-undef':                    [2],
-	'no-path-concat':              [2],
-	'no-proto':                    [2],
-	'no-redeclare':                [2],
-	'no-return-assign':            [2],
-	'no-self-compare':             [2],
-	'no-sequences':                [2],
-	'no-shadow':                   [2],
-	'no-shadow-restricted-names':  [2],
-	'no-sparse-arrays':            [2],
-	'no-throw-literal':            [2],
-	'no-unexpected-multiline':     [2],
-	'no-unneeded-ternary':         [2],
-	'no-unused-vars':              [2],
-	'no-use-before-define':        [2],
-	'no-useless-call':             [2],
-	'no-void':                     [2],
-	'no-with':                     [2],
-	'operator-linebreak':          [2, 'after'],
-	'operator-assignment':         [2],
-	'quote-props':                 [2, 'consistent-as-needed'],
-	'use-isnan':                   [2],
-	'valid-typeof':                [2],
-	'wrap-iife':                   [2]
+export default {
+	'no-constant-condition': ['error'],
+	'no-control-regex': ['error'],
+	'no-delete-var': ['error'],
+	'no-div-regex': ['error'],
+	'no-duplicate-case': ['error'],
+	'no-empty-character-class': ['error'],
+	'no-ex-assign': ['error'],
+	'no-extend-native': ['error'],
+	'no-extra-bind': ['error'],
+	'no-extra-boolean-cast': ['error'],
+	'no-extra-parens': ['error'],
+	'no-fallthrough': ['error'],
+	'no-floating-decimal': ['error'],
+	'no-func-assign': ['error'],
+	'no-implied-eval': ['error'],
+	'no-inner-declarations': ['error'],
+	'no-invalid-regexp': ['error'],
+	'no-iterator': ['error'],
+	'no-label-var': ['error'],
+	'no-labels': ['error'],
+	'no-multi-str': ['error'],
+	'no-multiple-empty-lines': ['error'],
+	'no-native-reassign': ['error'],
+	'no-negated-in-lhs': ['error'],
+	'no-nested-ternary': ['error'],
+	'no-new': ['error'],
+	'no-new-func': ['error'],
+	'no-new-object': ['error'],
+	'no-new-require': ['error'],
+	'no-new-wrappers': ['error'],
+	'no-obj-calls': ['error'],
+	'no-octal': ['error'],
+	'no-octal-escape': ['error'],
+	'no-undef': ['error'],
+	'no-path-concat': ['error'],
+	'no-proto': ['error'],
+	'no-redeclare': ['error'],
+	'no-return-assign': ['error'],
+	'no-self-compare': ['error'],
+	'no-sequences': ['error'],
+	'no-shadow': ['error'],
+	'no-shadow-restricted-names': ['error'],
+	'no-sparse-arrays': ['error'],
+	'no-throw-literal': ['error'],
+	'no-unexpected-multiline': ['error'],
+	'no-unneeded-ternary': ['error'],
+	'no-unused-vars': ['error'],
+	'no-use-before-define': ['error'],
+	'no-useless-call': ['error'],
+	'no-void': ['error'],
+	'no-with': ['error'],
+	'operator-linebreak': ['error', 'after'],
+	'operator-assignment': ['error'],
+	'quote-props': ['error', 'consistent-as-needed'],
+	'use-isnan': ['error'],
+	'valid-typeof': ['error'],
+	'wrap-iife': ['error']
 };
+
