@@ -2,19 +2,18 @@
  * spacing rules
  * @author sarkiroka on 2017.02.17.
  */
-module.exports = {
-	'block-spacing':               [1],
-	'computed-property-spacing':   [2],
-	'indent':                      [2, 'tab', {SwitchCase: 1}],
-	'keyword-spacing':             [2],
-	'no-irregular-whitespace':     [2],
-	'no-regex-spaces':             [2],
-	'no-spaced-func':              [2],
-	'no-trailing-spaces':          [2],
-	// 'object-curly-spacing':        [2, 'never'],
-	'space-before-function-paren': [1, {named: 'never'}],
-	'space-before-blocks':         [1],
-	'space-infix-ops':             [2],
-	'space-unary-ops':             [2],
-	'spaced-comment':              [2, 'always']
+export default {
+	'block-spacing': ['warn'],
+	'computed-property-spacing': ['error'],
+	'indent': ['error', 'tab', { SwitchCase: 1 }],
+	'keyword-spacing': ['error'],
+	'no-irregular-whitespace': ['error'],
+	'no-regex-spaces': ['error'],
+	'no-spaced-func': ['error'],
+	'no-trailing-spaces': ['error'],
+	'space-before-function-paren': ['warn', { named: 'never' }],
+	'space-before-blocks': ['warn'],
+	'space-infix-ops': ['error'],
+	'space-unary-ops': ['error'],
+	'spaced-comment': ['error', 'always']
 };
